refactor: remove duplicate usecase
