+++ domain/src/main/java/com/amsterdam/domain/useCase/game/guessByPoster/DoGuessMovieByPosterHintUseCase.kt
@@ -1,7 +1,7 @@
 package com.amsterdam.domain.useCase.game.guessByPoster
 
 import com.amsterdam.domain.exceptions.NotEnoughPointsException
-import com.amsterdam.domain.useCase.game.GetTotalUserPointsUseCase
+import com.amsterdam.domain.useCase.common.GetTotalUserPointsUseCase
 import com.amsterdam.domain.useCase.game.UpdateUserGamePointsUseCase
 import kotlinx.coroutines.flow.first
