+++ viewModel/src/main/java/com/amsterdam/viewmodel/letsPlay/LetsPlayViewModel.kt
@@ -2,7 +2,7 @@ package com.amsterdam.viewmodel.letsPlay
 
 import androidx.lifecycle.viewModelScope
 import com.amsterdam.domain.useCase.game.GetAvailableGamesUseCase
-import com.amsterdam.domain.useCase.game.GetTotalUserPointsUseCase
+import com.amsterdam.domain.useCase.common.GetTotalUserPointsUseCase
 import com.amsterdam.viewmodel.letsPlay.LetsPlayUiState.GameDifficultyUiState
 import com.amsterdam.viewmodel.letsPlay.LetsPlayUiState.GameUiState.GameTypeUiState
 import com.amsterdam.viewmodel.shared.BaseViewModel
